refactor(http-types): split config aggregate module
